[Modules] (Moved configs, db and migrations)
- Moved configs into appropriate modules.
- Moved gorm db setup into internal/db folder.
- Since each module should have its own table / repository, I moved migrations into a user module (since it is used there). Also updated docker-compose.yml file.
