+++ docker-compose.yml
@@ -24,7 +24,7 @@ services:
       - "8080:8080"
     env_file:
       - .env
-    command: /bin/sh -c "migrate -path pkg/common/db/migrations -database 'postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:${DB_PORT}/${DB_NAME}?sslmode=disable' up && ./main"
+    command: /bin/sh -c "migrate -path internal/module/user/migrations -database 'postgres://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:${DB_PORT}/${DB_NAME}?sslmode=disable' up && ./main"
 
 volumes:
   postgres_volume:
